+++ client-app/core/composables/useAuth.ts
@@ -1,24 +1,11 @@
-import { useLocalStorage, createGlobalState } from "@vueuse/core";
+import type { AfterFetchContext } from "@vueuse/core";
+import { createGlobalState, useLocalStorage } from "@vueuse/core";
 import { computed, ref } from "vue";
 import { useFetch } from "@/core/api/common";
 import { errorHandler, toServerError } from "@/core/api/common/utils";
 import { globals } from "@/core/globals";
 import { TabsType, unauthorizedErrorEvent, useBroadcast, userBeforeUnauthorizeEvent } from "@/shared/broadcast";
-import type { AfterFetchContext } from "@vueuse/core";
-
-type IdentityErrorType = {
-  code?: string;
-  description?: string;
-};
-
-type ConnectTokenResponseType = {
-  expires_in?: number;
-  access_token?: string;
-  refresh_token?: string;
-  errors?: Array<IdentityErrorType>;
-  error: string;
-  token_type?: string;
-};
+import type { ConnectTokenResponseType } from "../types";
 
 function _useAuth() {
   const broadcast = useBroadcast();
@@ -107,12 +94,10 @@ function _useAuth() {
   }
 
   async function externalSignInCallback(): Promise<void> {
-    const params = new URLSearchParams({
+    getTokenParams.value = new URLSearchParams({
       grant_type: "external_sign_in",
       scope: "offline_access",
     });
-
-    getTokenParams.value = params;
 
     await (getTokenRequest = getToken(true));
   }
