+++ client-app/core/composables/useImpersonate.ts
@@ -6,13 +6,7 @@ import { useAuth } from "@/core/composables/useAuth";
 import { Logger } from "@/core/utilities";
 import { TabsType, useBroadcast, reloadAndOpenMainPage } from "@/shared/broadcast";
 import { useNotifications } from "@/shared/notification";
-
-type ConnectTokenResponseType = {
-  expires_in: number;
-  access_token: string;
-  refresh_token: string;
-  token_type: string;
-};
+import type { ConnectTokenResponseType } from "../types";
 
 export function _useImpersonate() {
   const { setTokenType, setAccessToken, setExpiresAt, setRefreshToken } = useAuth();
@@ -39,20 +33,24 @@ export function _useImpersonate() {
       if (!data.value || error.value) {
         status.value = "error";
       } else {
-        const { access_token, token_type, expires_in, refresh_token } = data.value;
+        const { access_token, token_type, expires_in, refresh_token, errors } = data.value;
 
-        setAccessToken(access_token);
-        setExpiresAt(expires_in);
-        setTokenType(token_type);
-        setRefreshToken(refresh_token);
+        if (access_token && token_type && expires_in && refresh_token) {
+          setAccessToken(access_token);
+          setExpiresAt(expires_in);
+          setTokenType(token_type);
+          setRefreshToken(refresh_token);
+          status.value = "success";
+          notifications.success({ text: t("pages.account.impersonate.success") });
 
-        status.value = "success";
-        notifications.success({ text: t("pages.account.impersonate.success") });
-
-        // reload all tabs to renew state
-        setTimeout(() => {
-          void broadcast.emit(reloadAndOpenMainPage, null, TabsType.ALL);
-        }, 1000);
+          // reload all tabs to renew state
+          setTimeout(() => {
+            void broadcast.emit(reloadAndOpenMainPage, null, TabsType.ALL);
+          }, 1000);
+        } else {
+          Logger.error(impersonate.name, errors);
+          status.value = "error";
+        }
       }
     } catch (e) {
       Logger.error(impersonate.name, e);
